refactor(goal-achievement): extract utility functions
Split out utility functions from GoalAchievement component:
- Created utils/ directory with focused modules
- Extracted GOALS constant to goalConfig.ts
- Moved stats calculations to statsCalculators.ts
- Isolated gym session logic to gymSessionUtils.ts
- Added proper TypeScript interfaces in types.ts
- Updated main component to use new utilities

No functional changes, just better code organization.
