+++ src/components/dashboard/goal-achievement/utils/goalConfig.ts
@@ -1,0 +1,8 @@
+import type { GoalThresholds } from './types';
+
+export const GOALS: GoalThresholds = {
+  timeInOffice: 9,    // hours
+  calories: 2273,     // max calories
+  protein: 205,       // grams
+  weeklyGymSessions: 3, // sessions per week
+};
