+++ src/components/dashboard/goal-achievement/utils/types.ts
@@ -1,0 +1,25 @@
+import type { DailyEntry } from '@/types/dashboard';
+
+export interface GoalStats {
+  timeSuccess: number;
+  caloriesSuccess: number;
+  proteinSuccess: number;
+}
+
+export interface GoalStreaks {
+  timeStreak: number;
+  caloriesStreak: number;
+  proteinStreak: number;
+}
+
+export interface WeeklyGymStats {
+  currentWeekSessions: number;
+  weeklySuccessRate: number;
+}
+
+export interface GoalThresholds {
+  timeInOffice: number;  // hours
+  calories: number;      // max calories
+  protein: number;       // grams
+  weeklyGymSessions: number; // sessions per week
+}
